Modularize tracking add-on with adaptive marker detection
