+++ __init__.py
@@ -11,268 +11,15 @@ bl_info = {
     "category": "Tracking",
 }
 
-import bpy
-from bpy.types import Operator, Panel, PropertyGroup
-from bpy.props import BoolProperty, FloatProperty, IntProperty, PointerProperty
-
-
-class KAISERLICHSettings(PropertyGroup):
-    """Einstellungen für das Kaiserlich-Tracking."""
-
-    markers_per_frame: IntProperty(
-        name="Marker/Frame",
-        description="Gewünschte Anzahl Marker pro Frame",
-        default=10,
-        min=1,
-    )
-
-    min_track_length: IntProperty(
-        name="Frames/Track",
-        description="Minimale Track-Länge in Frames",
-        default=10,
-        min=1,
-    )
-
-    error_threshold: FloatProperty(
-        name="Error/Track",
-        description="Tolerierter Fehler pro Track",
-        default=1.0,
-        min=0.0,
-    )
-
-    auto_keyframes: BoolProperty(
-        name="Auto Keyframes",
-        description="Keyframes A und B automatisch bestimmen",
-        default=True,
-    )
-
-
-class KAISERLICH_PT_Panel(Panel):
-    """Panel für die Kaiserlich-Tracking Einstellungen."""
-
-    bl_label = "Kaiserlich"
-    bl_space_type = "CLIP_EDITOR"
-    bl_region_type = "UI"
-    bl_category = "Kaiserlich"
-    bl_context = "tracking"
-
-    def draw(self, context):
-        layout = self.layout
-        wm = context.window_manager
-        settings = wm.kaiserlich_settings
-
-        layout.prop(settings, "markers_per_frame")
-        layout.prop(settings, "min_track_length")
-        layout.prop(settings, "error_threshold")
-        layout.prop(settings, "auto_keyframes")
-        layout.operator("clip.kaiserlich_tracking", text="Track")
-        layout.separator()
-        layout.label(text="Diagnose")
-        layout.operator("clip.error_value", text="Error Value")
-
-
-class KAISERLICH_OT_Tracking(Operator):
-    """Führt das Kaiserlich-Tracking aus."""
-
-    bl_idname = "clip.kaiserlich_tracking"
-    bl_label = "Run Kaiserlich Tracking"
-    bl_description = (
-        "Automatisches Feature-Tracking und Kamera-Lösung basierend "
-        "auf den Kaiserlich-Parametern"
-    )
-    bl_options = {"REGISTER", "UNDO"}
-
-    def execute(self, context):
-        wm = context.window_manager
-        scene = context.scene
-        clip = getattr(context.space_data, "clip", None)
-
-        if clip is None:
-            self.report({"ERROR"}, "Kein aktiver Clip gefunden.")
-            return {"CANCELLED"}
-
-        markers_per_frame = wm.kaiserlich_settings.markers_per_frame
-        min_frames = wm.kaiserlich_settings.min_track_length
-        error_limit = wm.kaiserlich_settings.error_threshold
-
-        tracking = clip.tracking
-        settings = tracking.settings
-
-        tracking_obj = tracking.objects.active
-        if tracking_obj is None:
-            self.report({"WARNING"}, "Kein aktives Tracking-Objekt vorhanden.")
-            return {"CANCELLED"}
-        clip.use_proxy = False
-        clip.proxy.build_25 = False
-        clip.proxy.build_50 = False
-        clip.proxy.build_75 = False
-        clip.proxy.build_100 = False
-
-        image_width = float(clip.size[0])
-        settings.default_pattern_size = max(int(image_width / 100), 5)
-        settings.default_search_size = settings.default_pattern_size
-        settings.default_motion_model = "Loc"
-        settings.default_pattern_match = "KEYFRAME"
-        settings.default_correlation_min = 0.9
-
-        bpy.ops.clip.select_all(action="DESELECT")
-        margin = 16
-        threshold = 0.5
-        min_distance = int(image_width * 0.05)
-        bpy.ops.clip.detect_features(
-            placement="FRAME",
-            margin=margin,
-            threshold=threshold,
-            min_distance=min_distance,
-        )
-
-        tracks = tracking_obj.tracks
-        new_tracks = [t for t in tracks if t.select]
-        count_new = len(new_tracks)
-        if count_new < markers_per_frame:
-            self.report(
-                {"INFO"},
-                f"Nur {count_new} neue Marker gefunden, weniger als Ziel {markers_per_frame}.",
-            )
-
-        if not new_tracks:
-            self.report({"WARNING"}, "Keine Marker zum Tracken gefunden.")
-            return {"CANCELLED"}
-
-        for track in new_tracks:
-            track.motion_model = "Loc"
-            track.pattern_match = "KEYFRAME"
-            track.use_red_channel = True
-            track.use_green_channel = True
-            track.use_blue_channel = True
-            track.use_mask = False
-            track.correlation_min = settings.default_correlation_min
-
-        bpy.ops.clip.track_markers(backwards=False, sequence=True)
-        bpy.ops.clip.track_markers(backwards=True, sequence=True)
-
-        for track in list(tracks):
-            if len(track.markers) < min_frames:
-                track.select = True
-            else:
-                track.select = False
-        if any(t.select for t in tracks):
-            bpy.ops.clip.delete_track()
-
-        bpy.ops.clip.clean_tracks(frames=0, error=error_limit, action="DELETE_TRACK")
-
-        frame_start = int(scene.frame_start)
-        frame_end = int(scene.frame_end)
-        clip_start = getattr(clip, "frame_start", 0)
-        clip_end = clip_start + getattr(clip, "frame_duration", 0) - 1
-
-        if frame_start < clip_start or frame_end > clip_end:
-            self.report(
-                {"WARNING"}, "Scene-Frames liegen außerhalb des Clip-Zeitraums."
-            )
-            return {"CANCELLED"}
-
-        valid_tracks = [
-            t for t in tracking_obj.tracks if len([m for m in t.markers if not m.mute]) >= 5
-        ]
-        if len(valid_tracks) < 8:
-            self.report(
-                {"ERROR"},
-                "Zu wenige gültige Tracks für Kameralösung (mind. 8 benötigt)",
-            )
-            return {"CANCELLED"}
-
-        auto_keyframes = wm.kaiserlich_settings.auto_keyframes
-        if auto_keyframes:
-            bpy.ops.clip.set_keyframe_a()
-            bpy.ops.clip.set_keyframe_b()
-        else:
-            tracking_obj.keyframe_a = frame_start
-            tracking_obj.keyframe_b = frame_end
-
-        try:
-            bpy.ops.clip.solve_camera()
-        except RuntimeError as e:
-            self.report({"ERROR"}, f"Lösung fehlgeschlagen: {str(e)}")
-            return {"CANCELLED"}
-
-        solve_error = tracking_obj.reconstruction_error
-        self.report(
-            {"INFO"},
-            f"Kamera gelöst. Durchschnittlicher Solve-Fehler: {solve_error:.3f}",
-        )
-
-        bpy.ops.clip.set_scene_frames()
-        bpy.ops.clip.setup_tracking_scene()
-
-        for obj in scene.objects:
-            if obj.type == "CAMERA" and obj.constraints:
-                for constr in obj.constraints:
-                    if constr.type == "CAMERA_SOLVER":
-                        bpy.context.view_layer.objects.active = obj
-                        bpy.ops.clip.constraint_to_fcurve()
-                        constr.mute = True
-                        break
-
-        return {"FINISHED"}
-
-
-class CLIP_OT_error_value(Operator):
-    """Berechnet die durchschnittliche Standardabweichung der Marker-Positionen."""
-
-    bl_idname = "clip.error_value"
-    bl_label = "Compute Error Value"
-    bl_options = {"REGISTER"}
-
-    def execute(self, context):
-        clip = getattr(context.space_data, "clip", None)
-        if clip is None:
-            self.report({"ERROR"}, "Kein aktiver Clip gefunden.")
-            return {"CANCELLED"}
-
-        tracking = clip.tracking
-        tracking_obj = tracking.objects.active
-        if tracking_obj is None:
-            self.report({"ERROR"}, "Kein aktives Tracking-Objekt vorhanden.")
-            return {"CANCELLED"}
-
-        total_std = 0.0
-        count = 0
-        for t in tracking_obj.tracks:
-            positions = [m.co for m in t.markers if not m.mute]
-            if len(positions) < 2:
-                continue
-            mean_x = sum(p[0] for p in positions) / len(positions)
-            mean_y = sum(p[1] for p in positions) / len(positions)
-            variance = sum((p[0] - mean_x) ** 2 + (p[1] - mean_y) ** 2 for p in positions) / len(positions)
-            total_std += variance ** 0.5
-            count += 1
-
-        if count == 0:
-            self.report({"INFO"}, "Keine gültigen Tracks zur Berechnung.")
-            return {"CANCELLED"}
-
-        avg_std = total_std / count
-        self.report({"INFO"}, f"Durchschnittliche Std-Abweichung: {avg_std:.4f}")
-        return {"FINISHED"}
+from . import ui
 
 
 def register():
-    bpy.utils.register_class(KAISERLICHSettings)
-    bpy.utils.register_class(KAISERLICH_OT_Tracking)
-    bpy.utils.register_class(CLIP_OT_error_value)
-    bpy.utils.register_class(KAISERLICH_PT_Panel)
-    bpy.types.WindowManager.kaiserlich_settings = PointerProperty(
-        type=KAISERLICHSettings
-    )
+    ui.register()
 
 
 def unregister():
-    del bpy.types.WindowManager.kaiserlich_settings
-    bpy.utils.unregister_class(KAISERLICH_PT_Panel)
-    bpy.utils.unregister_class(CLIP_OT_error_value)
-    bpy.utils.unregister_class(KAISERLICH_OT_Tracking)
-    bpy.utils.unregister_class(KAISERLICHSettings)
+    ui.unregister()
 
 
 if __name__ == "__main__":
